feat: se mejoró la gestión de roles de usuario en Moodle
- Se trasladó la función `get_user_roles` al servicio `MoodleService` para centralizar la lógica de recuperación de roles.
- Se actualizó la llamada a `get_user_roles` en `ListAccountRequestsController` para utilizar el nuevo método del servicio.
